chore(fly): consolidate TigrisCredentialsMissing and FlyKmsError into Errors.ts
