+++ packages/alchemy/src/Fly/Decrypt.ts
@@ -1,5 +1,5 @@
 import type { DecryptSecretKeyError } from "@distilled.cloud/fly-io/machines";
-import type { FlyKmsError } from "./KmsError.ts";
+import type { FlyKmsError } from "./Errors.ts";
 import type * as Effect from "effect/Effect";
 import type * as Redacted from "effect/Redacted";
 import * as Binding from "../Binding.ts";
