+++ packages/alchemy/src/Fly/Verify.ts
@@ -1,5 +1,5 @@
 import type { VerifySecretKeyError } from "@distilled.cloud/fly-io/machines";
-import type { FlyKmsError } from "./KmsError.ts";
+import type { FlyKmsError } from "./Errors.ts";
 import type * as Effect from "effect/Effect";
 import * as Binding from "../Binding.ts";
 import type { RuntimeContext } from "../RuntimeContext.ts";
