+++ packages/alchemy/src/Fly/index.ts
@@ -1,7 +1,6 @@
 export * from "./App.ts";
 export * from "./AuthProvider.ts";
 export * from "./Bucket.ts";
-export * from "./BucketTypes.ts";
 export * from "./Catalog.ts";
 export * from "./DeleteObject.ts";
 export * from "./DeleteObjectHttp.ts";
@@ -14,6 +13,7 @@ export * from "./DecryptHttp.ts";
 export * from "./Encrypt.ts";
 export * from "./EncryptHttp.ts";
 export * from "./Environment.ts";
+export * from "./Errors.ts";
 export * from "./Exec.ts";
 export * from "./ExecHttp.ts";
 export * from "./GetObject.ts";
@@ -23,7 +23,6 @@ export * from "./GetSecretHttp.ts";
 export * from "./HeadObject.ts";
 export * from "./HeadObjectHttp.ts";
 export * from "./IpAssignment.ts";
-export * from "./KmsError.ts";
 export * from "./ListObjectsV2.ts";
 export * from "./ListObjectsV2Http.ts";
 export * from "./ListSecrets.ts";
